Update dependencies and enhance modal functionality
Upgrades various dependencies, including `expo-router` and `expo-constants`,
to their latest versions for improved compatibility and performance.

Introduces a modal in the Chat component with a slide animation and
visibility toggle for enhanced UI functionality. Updates the initial
route in the layout to 'index' for better navigation flow.

Adds new packages like `react-native-modal` and `@gorhom/bottom-sheet`
to expand UI capabilities and improve UX.